This merge request merges the changes from the ~"21.0" branch between [release/21.0.81](https://gitlab.cern.ch/atlas/athena/tags/release/21.0.81)
and [release/21.0.83](https://gitlab.cern.ch/atlas/athena/tags/release/21.0.83) into the ~"21.9" branch. This is done via a merge of
[release/21.3.7](https://gitlab.cern.ch/atlas/athena/tags/release/21.3.7).

[release/21.0.82](https://gitlab.cern.ch/atlas/athena/tags/release/21.0.82) is a Tier0 release. It's FT0-preserving.

[release/21.0.83](https://gitlab.cern.ch/atlas/athena/tags/release/21.0.83) is a Tier0 release. It's FT0-violating. It includes the FT0V MRs:
!14425: Sweeping !14379 from ~"21.2" to ~"21.0". Added flag for HI b-tagging reconstruction, with default to off


Former-commit-id: 31840453bf1704a63bdc6ceb48825a17d71ef4d6

+++ LArCalorimeter/LArCellRec/src/LArCellDeadOTXCorr.cxx
@@ -1092,7 +1092,7 @@ double LArCellDeadOTXCorr::getEnergyCalibration(double eta, int type, double ene
 
 double LArCellDeadOTXCorr::getMaxOverSumRatio(const std::vector<uint_least16_t>& ADCsamples, int pedestal)
 {
-	uint_least16_t max = -9999;
+	int max = -9999;
 	int sum = 0;
 	double ratio = -9999.;
 	unsigned int maxIndex = 0;
